+++ FlOpEDT/base/views.py
@@ -46,7 +46,7 @@ from django.views.generic import RedirectView
 from FlOpEDT.decorators import dept_admin_required, tutor_required
 from FlOpEDT.settings.base import COSMO_MODE
 
-from people.models import Tutor, UserDepartmentSettings, User
+from people.models import Tutor, UserDepartmentSettings, User, NotificationsPreferences
 
 from displayweb.admin import BreakingNewsResource
 from displayweb.models import BreakingNews
@@ -132,9 +132,6 @@ def edt(req, year=None, week=None, splash_id=0, **kwargs):
         copie = 0
         gp = ''
 
-    # une_salle = RoomGroup.objects.all()[1].name
-    une_salle = 'salle?'
-
     if req.user.is_authenticated:
         name_usr = req.user.username
         try:
@@ -151,7 +148,6 @@ def edt(req, year=None, week=None, splash_id=0, **kwargs):
                                 'week': week,
                                 'year': year,
                                 'promo': promo,
-                                'une_salle': une_salle,
                                 'copie': copie,
                                 'gp': gp,
                                 'name_usr': name_usr,
@@ -188,15 +184,12 @@ def edt_light(req, year=None, week=None, **kwargs):
         gp_w = 30
         svg_top_m = 40
 
-    une_salle = "salle?"  # RoomGroup.objects.all()[0].name
-
     return TemplateResponse(req, 'base/show-edt-light.html',
                             {
                                 'all_weeks': week_list(),
                                 'week': week,
                                 'year': year,
                                 'promo': promo,
-                                'une_salle': une_salle,
                                 'copie': 0,
                                 'gp': '',
                                 'name_usr': '',
@@ -227,6 +220,7 @@ def preferences(req, **kwargs):
 @login_required
 def stype(req, *args, **kwargs):
     err = ''
+    user_notifications_pref = queries.get_notification_preference(req.user)
     if req.method == 'GET':
         return TemplateResponse(req,
                                 'base/show-stype.html',
@@ -235,6 +229,7 @@ def stype(req, *args, **kwargs):
                                  'name_usr': req.user.username,
                                  'usr_pref_hours': req.user.tutor.pref_hours_per_day,
                                  'usr_max_hours': req.user.tutor.max_hours_per_day,
+                                 'user_notifications_pref': user_notifications_pref,
                                  'err': err,
                                  'current_year': current_year,
                                  'time_settings': queries.get_time_settings(req.department),
@@ -270,6 +265,7 @@ def stype(req, *args, **kwargs):
                                  'name_usr': req.user.username,
                                  'usr_pref_hours': req.user.tutor.pref_hours_per_day,
                                  'usr_max_hours': req.user.tutor.max_hours_per_day,
+                                 'user_notifications_pref': user_notifications_pref,
                                  'err': err,
                                  'current_year': current_year,
                                  'time_settings': queries.get_time_settings(req.department),
@@ -278,9 +274,6 @@ def stype(req, *args, **kwargs):
 
 @tutor_required
 def room_preference(req, department, tutor=None):
-    
-
-    
     roomtypes = RoomType.objects.filter(department=req.department)\
                                 .prefetch_related('members')
     roomgroups = Room.objects.filter(types__in=roomtypes)
@@ -355,8 +348,7 @@ def user_perfect_day_changes(req, username=None, *args, **kwargs):
         t.pref_hours_per_day = user_pref_hours
         t.max_hours_per_day = user_max_hours
         t.save()
-    return redirect('base:stype', req.department)
-
+    return redirect('base:preferences', req.department)
 
 @login_required
 def fetch_perfect_day(req, username=None, *args, **kwargs):
@@ -366,6 +358,28 @@ def fetch_perfect_day(req, username=None, *args, **kwargs):
         perfect_day['pref'] = t.pref_hours_per_day
         perfect_day['max'] = t.max_hours_per_day
     return JsonResponse(perfect_day, safe=False)
+
+
+@login_required
+def fetch_user_notifications_pref(req, username=None, *args, **kwargs):
+    try:
+        user = User.objects.get(username=username)
+    except User.DoesNotExist:
+        user = None
+    
+    return JsonResponse({"nb_weeks": queries.get_notification_preference(user)},
+                        safe=False)
+
+@login_required
+def user_notifications_pref_changes(req, username=None, *args, **kwargs):
+    if username is not None:
+        u = User.objects.get(username=username)
+        n, created = NotificationsPreferences.objects.get_or_create(user=u)
+        data = req.POST
+        user_notifications_pref = int(data['user_notifications_pref'])
+        n.nb_of_notified_weeks = user_notifications_pref
+        n.save()
+    return redirect('base:preferences', req.department)
 
 
 @dept_admin_required
@@ -1043,6 +1057,7 @@ def clean_change(year, week, old_version, change, work_copy=0, initiator=None, a
     tutor_old = sched_course.tutor
     if tutor_old is None:
         tutor_old = sched_course.course.tutor
+        sched_course.tutor = tutor_old
         
     course_log = CourseModification(course=course,
                                     old_week=week,
@@ -1060,14 +1075,11 @@ def clean_change(year, week, old_version, change, work_copy=0, initiator=None, a
 
     # Rooms
     try:
-        new_room = Room.objects.get(name=change['room'])
+        new_room = Room.objects.get(name=change['room']) \
+            if change['room'] != '' else None
         ret['sched'].room = new_room
-    except ObjectDoesNotExist:
-        if new_room == 'salle?' or new_room is None:
-            raise Exception('Oublié de trouver une salle ' \
-                  'pour un cours ?')
-        else:
-            raise Exception(f"Problème : salle {change['room']} inconnue")
+    except Room.DoesNotExist:
+        raise Exception(f"Problème : salle {change['room']} inconnue")
 
     # Timing
     ret['sched'].start_time = change['start']